敵キャラクター、敵工場、衝突判定空間、スケジューラー、マップ、敵の出撃をリファクタリング 原田

+++ Program/SnowDrop/Classes/Data/LaunchTrigger/LaunchTrigger.cpp
@@ -11,45 +11,43 @@
 //=======================================
 //デストラクタ
 CEnemyLaunchTrigger::~CEnemyLaunchTrigger() {
+	//敵出撃データを削除
 	SAFE_DELETE(this->m_pLaunchdata);
 }
 
 /**
-* @desc		イベントを行えるかどうか
-* @retrun	true...イベント実行可
-*/
+ * @desc	イベントを行えるかどうか
+ * @retrun	true...イベント実行可
+ */
 bool CEnemyLaunchTrigger::isTrigger(){
+	//敵出撃データに中身がなければ
 	if(this->m_pLaunchdata == NULL){
+		//イベント実行　不可　を返す。
 		return false;
 	}
+	//イベント実行　可　を返す
 	return true;
 }
 
 /**
-* @desc		設定されている敵出撃データを元に敵を生成する
-*			トリガーを元にイベントを実行
-*/
+ * @desc	設定されている敵出撃データを元に敵を生成する
+ * @tips	トリガーを元にイベントを実行
+ */
 CCharacter* CEnemyLaunchTrigger::action(){
 
 	//敵を生成
+	//引数に出撃させるポイントと敵のタイプを渡す
 	CCharacter* pEnemyCharacter = (CCharacter*)CEnemyFactoryManager::getInstance()->create(
 		this->m_pLaunchdata->m_pos,
-		(int)this->m_pLaunchdata->m_type
+		this->m_pLaunchdata->m_type
 	);
 
 	//発射し終わったかどうかのフラグを立てる
 	this->m_activeFlag = false;
 
+	//生成した敵キャラクターを返す。
 	return pEnemyCharacter;
 }
-
-//=======================================
-//
-//  弾発射トリガー
-//		弾発射データをトリガーとして弾発射というイベントを実行させる
-//
-//=======================================
-
 
 //=======================================
 //
@@ -58,27 +56,32 @@ CCharacter* CEnemyLaunchTrigger::action(){
 //=======================================
 //デストラクタ
 CGimmickLaunchTrigger::~CGimmickLaunchTrigger() {
+	//ギミック出撃データを削除
 	SAFE_DELETE(this->m_pLaunchdata);
 }
 
 /**
-* @desc		イベントを行えるかどうか
-* @retrun	true...イベント実行可
-*/
+ * @desc	イベントを行えるかどうか
+ * @retrun	true...イベント実行可
+ */
 bool CGimmickLaunchTrigger::isTrigger() {
+	//ギミック出撃データに中身がなければ
 	if (this->m_pLaunchdata == NULL) {
+		//イベント実行　不可　を返す
 		return false;
 	}
+	//イベント実行　可　を返す
 	return true;
 }
 
 /**
-* @desc		設定されている敵出撃データを元に敵を生成する
-*			トリガーを元にイベントを実行
-*/
+ * @desc	設定されている敵出撃データを元に敵を生成する
+ * @tips	トリガーを元にイベントを実行
+ */
 CCharacter* CGimmickLaunchTrigger::action() {
 
-	//弾を生成
+	//ギミックを生成
+	//引数にギミックのタイプと出撃させる位置を渡す
 	CCharacter* pGimmickCharacter = (CCharacter*)CGimmickFactoryManager::getInstance()->create(
 		(int)this->m_pLaunchdata->m_type,
 		this->m_pLaunchdata->m_pos
@@ -87,6 +90,7 @@ CCharacter* CGimmickLaunchTrigger::action() {
 	//発射し終わったかどうかのフラグを立てる
 	this->m_activeFlag = false;
 
+	//生成したギミックを返す。
 	return pGimmickCharacter;
 }
 
@@ -105,38 +109,43 @@ CLaunchScheduler::CLaunchScheduler(){}
 //共有インスタンスの取得
 CLaunchScheduler* CLaunchScheduler::getInstance() {
 
+	//共有インスタンスが存在しなければ
 	if (CLaunchScheduler::m_pSharedScheduler == NULL) {
+		//共有インスタンスを生成
 		CLaunchScheduler::m_pSharedScheduler = new CLaunchScheduler();
 	}
+	//共有インスタンスを返す。
 	return CLaunchScheduler::m_pSharedScheduler;
 }
 
 //共有インスタンスの破棄
 void CLaunchScheduler::removeInstance() {
+	//共有インスタンスの削除
 	SAFE_DELETE(CLaunchScheduler::m_pSharedScheduler);
 }
 
 //デストラクタ
 CLaunchScheduler::~CLaunchScheduler() {
+	//共有インスタンスの削除
 	SAFE_DELETE(this->m_pLauncher);
 }
 
 /**
-* @desc		発射台の生成
-* @param	発射台に設定する出撃スケジュールのアドレス
-*/
+ * @desc	発射台の生成
+ * @param	発射台に設定する出撃スケジュールのアドレス
+ */
 void CLaunchScheduler::createLauncher(std::vector<CLaunchTrigger*>* pLaunchSchedule) {
-
+	//発射台が存在しなければ
 	if (this->m_pLauncher == NULL) {
+		//発射台を生成
 		this->m_pLauncher = new CLauncher(pLaunchSchedule);
 	}
-
 }
 
 /**
-* @desc		キャラクターの出撃
-* @param	取り付けるレイヤー
-*/
+ * @desc	キャラクターの出撃
+ * @param	取り付けるレイヤー
+ */
 void CLaunchScheduler::launchCharacter(cocos2d::Layer* pLayer) {
 
 	//スケジューラに取り付けられている起動ができるトリガー全てを起動する
@@ -146,7 +155,7 @@ void CLaunchScheduler::launchCharacter(cocos2d::Layer* pLayer) {
 		//起動可能な出撃トリガーを取得して出撃トリガーイテレーターを次へ進める
 		CLaunchTrigger* pTrigger = itr->next();
 
-		//出撃トリガーを起動
+		//出撃トリガーを起動してキャラクターを生成し取得
 		CCharacter* pChara = pTrigger->action();
 
 		//NULLチェック
@@ -158,8 +167,6 @@ void CLaunchScheduler::launchCharacter(cocos2d::Layer* pLayer) {
 			//キャラクターをメインレイヤーに取り付ける
 			pLayer->addChild(pChara);
 		}
-
 	}
-
 }
